Changed SlidingTabs to expand to fill and fixed tablet main activity split color theming
I had to manually override the listView and seekBar styles to red for
the Mood/Manual fragment because it is not possible to add themes to
fragments. Changing tabs now use layout weight to expand to fill, so I
reduced the tab indicator height to visually balance their new,
expanding width.